Api update Movie entity
VoteAverage now a double instead of int

+++ MovieToGoAPI/MovieToGoAPI/DTOs/Movies/MovieDTO.cs
@@ -21,7 +21,7 @@ namespace MovieToGoAPI.DTOs.Movies
 
         public int TheMovieDbId { get; set; }
 
-        public int? VoteAverage { get; set; }
+        public double? VoteAverage { get; set; }
 
         public int? VoteCount { get; set; }
